Edit JavaDoc, remove during joiner, move IntervalJoinerTests to JoinerTests
