+++ optaplanner-core/src/main/java/org/optaplanner/core/api/score/stream/Joiners.java
@@ -152,97 +152,46 @@ public final class Joiners {
      */
 
     /**
-     * For pairs of intervals [a,b) and [c,d), returns only those pairs where [a,b) and [c,d) overlap (or, in other
-     * words, there exists an x such that x in [a,b) and x in [c,d)).
-     * These are exactly the pairs where {@code a < d} and {@code b > c} (or in other words, when the first interval
-     * starts before the second one ends and the first interval ends after the second interval begins).
+     * Joins every A and B that overlap for an interval which is specified by a start and end property on both A and B.
+     * These are exactly the pairs where {@code A.start < B.end} and {@code A.end > B.start}.
+     * 
+     * @param startMapping maps the argument to the start point of its interval (inclusive)
+     * @param endMapping maps the argument to the end point of its interval (exclusive)
+     * 
+     * @param <A> the type of both the first and second argument
+     * @param <Property_> the type used to define the interval, comparable
+     * 
+     * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B) where
+     *         A's and B's intervals (as defined by the function mapping) overlap
+     */
+    public static <A, Property_ extends Comparable<Property_>> BiJoiner<A, A> overlapping(
+            Function<A, Property_> startMapping,
+            Function<A, Property_> endMapping) {
+        return overlapping(startMapping, endMapping, startMapping, endMapping);
+    }
+
+    /**
+     * As defined by {@link #overlapping(Function, Function)}.
+     * 
+     * @param leftStartMapping maps the first argument to its interval start point (inclusive)
+     * @param leftEndMapping maps the first argument to its interval end point (exclusive)
+     * @param rightStartMapping maps the second argument to its interval start point (inclusive)
+     * @param rightEndMapping maps the second argument to its interval end point (exclusive)
      * 
      * @param <A> the type of the first argument
      * @param <B> the type of the second argument
      * @param <Property_> the type used to define the interval, comparable
      * 
-     * @param leftStartMapping maps the first argument to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first argument to the first interval end point (exclusive)
-     * @param rightStartMapping maps the second argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the second argument to the second interval end point (exclusive)
-     * 
      * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B) where
      *         A's and B's intervals (as defined by the function mapping) overlap
      */
-    public static <A, B, Property_ extends Comparable<Property_>> BiJoiner<A, B> overlaps(
+    public static <A, B, Property_ extends Comparable<Property_>> BiJoiner<A, B> overlapping(
             Function<A, Property_> leftStartMapping,
             Function<A, Property_> leftEndMapping,
             Function<B, Property_> rightStartMapping,
             Function<B, Property_> rightEndMapping) {
         return AbstractBiJoiner.merge(Joiners.lessThan(leftStartMapping, rightEndMapping),
                 Joiners.greaterThan(leftEndMapping, rightStartMapping));
-    }
-
-    /**
-     * Syntactic sugar for {@link #overlaps(Function, Function, Function, Function)} where
-     * both elements of the tuple (A,B) are the same type and have the same interval
-     * defining function. Identical to
-     * {@code overlaps(startMapping, endMapping, startMapping, endMapping)}.
-     * 
-     * @param <A> the type of both the first and second argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param startMapping maps the argument to the start point of its interval (inclusive)
-     * @param endMapping maps the argument to the end point of its interval (exclusive)
-     * 
-     * @return see {@link #overlaps(Function, Function, Function, Function)}
-     */
-    public static <A, Property_ extends Comparable<Property_>> BiJoiner<A, A> overlaps(
-            Function<A, Property_> startMapping,
-            Function<A, Property_> endMapping) {
-        return overlaps(startMapping, endMapping, startMapping, endMapping);
-    }
-
-    /**
-     * For pairs of intervals [a,b) and [c,d), returns only those pairs where [a,b) contains [c,d) (or, in other
-     * words, for any x in [c, d), x is also in [a,b)).
-     * These are exactly the pairs where {@code a <= c} and {@code b >= d} (or in other words, when the first interval
-     * starts before the second one starts and the first interval ends after the second interval ends).
-     * 
-     * @param <A> the type of the first argument
-     * @param <B> the type of the second argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param leftStartMapping maps the first argument to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first argument to the first interval end point (exclusive)
-     * @param rightStartMapping maps the second argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the second argument to the second interval end point (exclusive)
-     * 
-     * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B) where
-     *         A's interval contains B's interval (as defined by the function mapping)
-     */
-    public static <A, B, Property_ extends Comparable<Property_>> BiJoiner<A, B> during(
-            Function<A, Property_> leftStartMapping,
-            Function<A, Property_> leftEndMapping,
-            Function<B, Property_> rightStartMapping,
-            Function<B, Property_> rightEndMapping) {
-        return AbstractBiJoiner.merge(Joiners.lessThanOrEqual(leftStartMapping, rightStartMapping),
-                Joiners.greaterThanOrEqual(leftEndMapping, rightEndMapping));
-    }
-
-    /**
-     * Syntactic sugar for {@link #during(Function, Function, Function, Function)} where
-     * both elements of the tuple (A,B) are the same type and have the same interval
-     * defining function. Identical to
-     * {@code during(startMapping, endMapping, startMapping, endMapping)}.
-     * 
-     * @param <A> the type of both the first and second argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param startMapping maps the argument to the start point of its interval (inclusive)
-     * @param endMapping maps the argument to the end point of its interval (exclusive)
-     * 
-     * @return see {@link #during(Function, Function, Function, Function)}
-     */
-    public static <A, Property_ extends Comparable<Property_>> BiJoiner<A, A> during(
-            Function<A, Property_> startMapping,
-            Function<A, Property_> endMapping) {
-        return during(startMapping, endMapping, startMapping, endMapping);
     }
 
     // ************************************************************************
@@ -279,53 +228,28 @@ public final class Joiners {
     }
 
     /**
-     * Like {@link #overlaps(Function, Function, Function, Function)}, but for TriConstraintStream.
+     * As defined by {@link #overlapping(Function, Function)}.
+     * 
+     * @param leftStartMapping maps the first and second arguments to their interval start point (inclusive)
+     * @param leftEndMapping maps the first and second arguments to their interval end point (exclusive)
+     * @param rightStartMapping maps the third argument to its interval start point (inclusive)
+     * @param rightEndMapping maps the third argument to its interval end point (exclusive)
      * 
      * @param <A> the type of the first argument
      * @param <B> the type of the second argument
      * @param <C> the type of the third argument
      * @param <Property_> the type used to define the interval, comparable
      * 
-     * @param leftStartMapping maps the first and second arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first and second arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the third argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the third argument to the second interval end point (exclusive)
-     * 
      * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B,C) where
      *         (A,B)'s and C's intervals (as defined by the function mapping) overlap
      */
-    public static <A, B, C, Property_ extends Comparable<Property_>> TriJoiner<A, B, C> overlaps(
+    public static <A, B, C, Property_ extends Comparable<Property_>> TriJoiner<A, B, C> overlapping(
             BiFunction<A, B, Property_> leftStartMapping,
             BiFunction<A, B, Property_> leftEndMapping,
             Function<C, Property_> rightStartMapping,
             Function<C, Property_> rightEndMapping) {
         return AbstractTriJoiner.merge(Joiners.lessThan(leftStartMapping, rightEndMapping),
                 Joiners.greaterThan(leftEndMapping, rightStartMapping));
-    }
-
-    /**
-     * Like {@link #during(Function, Function, Function, Function)}, but for TriConstraintStream.
-     * 
-     * @param <A> the type of the first argument
-     * @param <B> the type of the second argument
-     * @param <C> the type of the third argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param leftStartMapping maps the first and second arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first and second arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the third argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the third argument to the second interval end point (exclusive)
-     * 
-     * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B,C) where
-     *         (A,B)'s interval contains C's interval (as defined by the function mapping)
-     */
-    public static <A, B, C, Property_ extends Comparable<Property_>> TriJoiner<A, B, C> during(
-            BiFunction<A, B, Property_> leftStartMapping,
-            BiFunction<A, B, Property_> leftEndMapping,
-            Function<C, Property_> rightStartMapping,
-            Function<C, Property_> rightEndMapping) {
-        return AbstractTriJoiner.merge(Joiners.lessThanOrEqual(leftStartMapping, rightStartMapping),
-                Joiners.greaterThanOrEqual(leftEndMapping, rightEndMapping));
     }
 
     // ************************************************************************
@@ -362,7 +286,12 @@ public final class Joiners {
     }
 
     /**
-     * Like {@link #overlaps(Function, Function, Function, Function)}, but for QuadConstraintStream.
+     * As defined by {@link #overlapping(Function, Function)}.
+     * 
+     * @param leftStartMapping maps the first, second and third arguments to their interval start point (inclusive)
+     * @param leftEndMapping maps the first, second and third arguments to their interval end point (exclusive)
+     * @param rightStartMapping maps the fourth argument to its interval start point (inclusive)
+     * @param rightEndMapping maps the fourth argument to its interval end point (exclusive)
      * 
      * @param <A> the type of the first argument
      * @param <B> the type of the second argument
@@ -370,47 +299,16 @@ public final class Joiners {
      * @param <D> the type of the fourth argument
      * @param <Property_> the type used to define the interval, comparable
      * 
-     * @param leftStartMapping maps the first, second and third arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first, second and third arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the fourth argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the fourth argument to the second interval end point (exclusive)
-     * 
      * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B,C,D)
      *         where (A,B,C)'s and D's intervals (as defined by the function mapping) overlap
      */
-    public static <A, B, C, D, Property_ extends Comparable<Property_>> QuadJoiner<A, B, C, D> overlaps(
+    public static <A, B, C, D, Property_ extends Comparable<Property_>> QuadJoiner<A, B, C, D> overlapping(
             TriFunction<A, B, C, Property_> leftStartMapping,
             TriFunction<A, B, C, Property_> leftEndMapping,
             Function<D, Property_> rightStartMapping,
             Function<D, Property_> rightEndMapping) {
         return AbstractQuadJoiner.merge(Joiners.lessThan(leftStartMapping, rightEndMapping),
                 Joiners.greaterThan(leftEndMapping, rightStartMapping));
-    }
-
-    /**
-     * Like {@link #during(Function, Function, Function, Function)}, but for QuadConstraintStream.
-     * 
-     * @param <A> the type of the first argument
-     * @param <B> the type of the second argument
-     * @param <C> the type of the third argument
-     * @param <D> the type of the fourth argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param leftStartMapping maps the first, second and third arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first, second and third arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the fourth argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the fourth argument to the second interval end point (exclusive)
-     * 
-     * @return nerver null, an indexed joiner that filters the constraint stream to only include elements (A,B,C,D)
-     *         where (A,B,C)'s interval contains D's interval (as defined by the function mapping)
-     */
-    public static <A, B, C, D, Property_ extends Comparable<Property_>> QuadJoiner<A, B, C, D> during(
-            TriFunction<A, B, C, Property_> leftStartMapping,
-            TriFunction<A, B, C, Property_> leftEndMapping,
-            Function<D, Property_> rightStartMapping,
-            Function<D, Property_> rightEndMapping) {
-        return AbstractQuadJoiner.merge(Joiners.lessThanOrEqual(leftStartMapping, rightStartMapping),
-                Joiners.greaterThanOrEqual(leftEndMapping, rightEndMapping));
     }
 
     // ************************************************************************
@@ -447,7 +345,12 @@ public final class Joiners {
     }
 
     /**
-     * Like {@link #overlaps(Function, Function, Function, Function)}, but for PentaConstraintStream.
+     * As defined by {@link #overlapping(Function, Function)}.
+     * 
+     * @param leftStartMapping maps the first, second, third and fourth arguments to their interval start point (inclusive)
+     * @param leftEndMapping maps the first, second, third and fourth arguments to their interval end point (exclusive)
+     * @param rightStartMapping maps the fifth argument to its interval start point (inclusive)
+     * @param rightEndMapping maps the fifth argument to its interval end point (exclusive)
      * 
      * @param <A> the type of the first argument
      * @param <B> the type of the second argument
@@ -456,48 +359,16 @@ public final class Joiners {
      * @param <E> the type of the fifth argument
      * @param <Property_> the type used to define the interval, comparable
      * 
-     * @param leftStartMapping maps the first, second, third and fourth arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first, second, third and fourth arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the fifth argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the fifth argument to the second interval end point (exclusive)
-     * 
      * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B,C,D,E)
      *         where (A,B,C,D)'s and E's intervals (as defined by the function mapping) overlap
      */
-    public static <A, B, C, D, E, Property_ extends Comparable<Property_>> PentaJoiner<A, B, C, D, E> overlaps(
+    public static <A, B, C, D, E, Property_ extends Comparable<Property_>> PentaJoiner<A, B, C, D, E> overlapping(
             QuadFunction<A, B, C, D, Property_> leftStartMapping,
             QuadFunction<A, B, C, D, Property_> leftEndMapping,
             Function<E, Property_> rightStartMapping,
             Function<E, Property_> rightEndMapping) {
         return AbstractPentaJoiner.merge(Joiners.lessThan(leftStartMapping, rightEndMapping),
                 Joiners.greaterThan(leftEndMapping, rightStartMapping));
-    }
-
-    /**
-     * Like {@link #during(Function, Function, Function, Function)}, but for PentaConstraintStream.
-     * 
-     * @param <A> the type of the first argument
-     * @param <B> the type of the second argument
-     * @param <C> the type of the third argument
-     * @param <D> the type of the fourth argument
-     * @param <E> the type of the fifth argument
-     * @param <Property_> the type used to define the interval, comparable
-     * 
-     * @param leftStartMapping maps the first, second, third and fourth arguments to the first interval start point (inclusive)
-     * @param leftEndMapping maps the first, second, third and fourth arguments to the first interval end point (exclusive)
-     * @param rightStartMapping maps the fifth argument to the second interval start point (inclusive)
-     * @param rightEndMapping maps the fifth argument to the second interval end point (exclusive)
-     * 
-     * @return never null, an indexed joiner that filters the constraint stream to only include elements (A,B,C,D,E)
-     *         where (A,B,C,D)'s interval contains E's interval (as defined by the function mapping)
-     */
-    public static <A, B, C, D, E, Property_ extends Comparable<Property_>> PentaJoiner<A, B, C, D, E> during(
-            QuadFunction<A, B, C, D, Property_> leftStartMapping,
-            QuadFunction<A, B, C, D, Property_> leftEndMapping,
-            Function<E, Property_> rightStartMapping,
-            Function<E, Property_> rightEndMapping) {
-        return AbstractPentaJoiner.merge(Joiners.lessThanOrEqual(leftStartMapping, rightStartMapping),
-                Joiners.greaterThanOrEqual(leftEndMapping, rightEndMapping));
     }
 
     private Joiners() {
