+++ optaplanner-core/src/test/java/org/optaplanner/core/api/score/stream/JoinersTest.java
@@ -317,4 +317,483 @@ public class JoinersTest {
         });
     }
 
+    private static final class Interval<A, B> {
+        final A start;
+        final A end;
+        final Function<A, B> map;
+
+        public Interval(A start, A end, Function<A, B> map) {
+            this.start = start;
+            this.end = end;
+            this.map = map;
+        }
+
+        public B getStart() {
+            return map.apply(start);
+        }
+
+        public B getEnd() {
+            return map.apply(end);
+        }
+
+        @Override
+        public String toString() {
+            return "(" + start.toString() + ", " + end.toString() + ")";
+        }
+
+        public static Interval<Long, BigDecimal> ofBigDecimal(Long start, Long end) {
+            return new Interval<>(start, end, BigDecimal::valueOf);
+        }
+
+        public static Interval<Long, BigInteger> ofBigInt(Long start, Long end) {
+            return new Interval<>(start, end, BigInteger::valueOf);
+        }
+    }
+
+    @Test
+    public void overlapsBiDifferentTypes() {
+        Function<Interval<Long, BigInteger>, Long> leftStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigInteger>, Long> leftEndMapping = interval -> interval.getEnd().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightEndMapping = interval -> interval.getEnd().longValue();
+
+        AbstractBiJoiner<Interval<Long, BigInteger>, Interval<Long, BigDecimal>> joiner =
+                (AbstractBiJoiner<Interval<Long, BigInteger>, Interval<Long, BigDecimal>>) Joiners.overlapping(leftStartMapping,
+                        leftEndMapping, rightStartMapping, rightEndMapping);
+
+        assertSoftly(softly -> {
+            // True cases (equals, overlaps, contains, starts, ends)
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("Case A = B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("B starts before A, A ends after B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A starts before B, B ends after A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 5L),
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B contains A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigDecimal(3L, 5L)))
+                    .as("A contains B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("A started by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B started by A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A ended by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B ended by A").isTrue();
+
+            // False Cases (before, after, meets)
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigDecimal(5L, 7L)))
+                    .as("A before B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(5L, 7L),
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B before A").isFalse();
+
+            // This is false since typically, when overlaps is used,
+            // end is exclusive, and start is inclusive,
+            // so 0-5, 5-10 do not overlap
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A meets B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B meets A").isFalse();
+        });
+    }
+
+    @Test
+    public void overlapsBiSameTypes() {
+        Function<Interval<Long, BigInteger>, Long> leftStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigInteger>, Long> leftEndMapping = interval -> interval.getEnd().longValue();
+
+        AbstractBiJoiner<Interval<Long, BigInteger>, Interval<Long, BigInteger>> joiner =
+                (AbstractBiJoiner<Interval<Long, BigInteger>, Interval<Long, BigInteger>>) Joiners.overlapping(leftStartMapping,
+                        leftEndMapping);
+
+        assertSoftly(softly -> {
+            // True cases (equals, overlaps, contains, starts, ends)
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    Interval.ofBigInt(1L, 5L)))
+                    .as("Case A = B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigInt(1L, 5L)))
+                    .as("B starts before A, A ends after B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    Interval.ofBigInt(3L, 7L)))
+                    .as("A starts before B, B ends after A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 5L),
+                    Interval.ofBigInt(1L, 7L)))
+                    .as("B contains A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigInt(3L, 5L)))
+                    .as("A contains B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigInt(1L, 3L)))
+                    .as("A started by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigInt(1L, 7L)))
+                    .as("B started by A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    Interval.ofBigInt(3L, 7L)))
+                    .as("A ended by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigInt(1L, 7L)))
+                    .as("B ended by A").isTrue();
+
+            // False Cases (before, after, meets)
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigInt(5L, 7L)))
+                    .as("A before B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(5L, 7L),
+                    Interval.ofBigInt(1L, 3L)))
+                    .as("B before A").isFalse();
+
+            // This is false since typically, when overlaps is used,
+            // end is exclusive, and start is inclusive,
+            // so 0-5, 5-10 do not overlap
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    Interval.ofBigInt(3L, 7L)))
+                    .as("A meets B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    Interval.ofBigInt(1L, 3L)))
+                    .as("B meets A").isFalse();
+        });
+    }
+
+    @Test
+    public void overlapsTri() {
+        BiFunction<Interval<Long, BigInteger>, String, Long> leftStartMapping =
+                (interval, ignored) -> interval.getStart().longValue();
+        BiFunction<Interval<Long, BigInteger>, String, Long> leftEndMapping =
+                (interval, ignored) -> interval.getEnd().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightEndMapping = interval -> interval.getEnd().longValue();
+
+        AbstractTriJoiner<Interval<Long, BigInteger>, String, Interval<Long, BigDecimal>> joiner =
+                (AbstractTriJoiner<Interval<Long, BigInteger>, String, Interval<Long, BigDecimal>>) Joiners.overlapping(
+                        leftStartMapping,
+                        leftEndMapping, rightStartMapping, rightEndMapping);
+
+        assertSoftly(softly -> {
+            // True cases (equals, overlaps, contains, starts, ends)
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("Case A = B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("B starts before A, A ends after B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A starts before B, B ends after A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 5L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B contains A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 5L)))
+                    .as("A contains B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("A started by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B started by A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A ended by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B ended by A").isTrue();
+
+            // False Cases (before, after, meets)
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(5L, 7L)))
+                    .as("A before B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(5L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B before A").isFalse();
+
+            // This is false since typically, when overlaps is used,
+            // end is exclusive, and start is inclusive,
+            // so 0-5, 5-10 do not overlap
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A meets B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B meets A").isFalse();
+        });
+    }
+
+    @Test
+    public void overlapsQuad() {
+        TriFunction<Interval<Long, BigInteger>, String, String, Long> leftStartMapping =
+                (interval, ignored1, ignored2) -> interval.getStart().longValue();
+        TriFunction<Interval<Long, BigInteger>, String, String, Long> leftEndMapping =
+                (interval, ignored1, ignored2) -> interval.getEnd().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightEndMapping = interval -> interval.getEnd().longValue();
+
+        AbstractQuadJoiner<Interval<Long, BigInteger>, String, String, Interval<Long, BigDecimal>> joiner =
+                (AbstractQuadJoiner<Interval<Long, BigInteger>, String, String, Interval<Long, BigDecimal>>) Joiners
+                        .overlapping(
+                                leftStartMapping,
+                                leftEndMapping, rightStartMapping, rightEndMapping);
+
+        assertSoftly(softly -> {
+            // True cases (equals, overlaps, contains, starts, ends)
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("Case A = B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("B starts before A, A ends after B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A starts before B, B ends after A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B contains A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 5L)))
+                    .as("A contains B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("A started by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B started by A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A ended by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B ended by A").isTrue();
+
+            // False Cases (before, after, meets)
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(5L, 7L)))
+                    .as("A before B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(5L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B before A").isFalse();
+
+            // This is false since typically, when overlaps is used,
+            // end is exclusive, and start is inclusive,
+            // so 0-5, 5-10 do not overlap
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A meets B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B meets A").isFalse();
+        });
+    }
+
+    @Test
+    public void overlapsPenta() {
+        QuadFunction<Interval<Long, BigInteger>, String, String, String, Long> leftStartMapping =
+                (interval, ignored1, ignored2, ignored3) -> interval.getStart().longValue();
+        QuadFunction<Interval<Long, BigInteger>, String, String, String, Long> leftEndMapping =
+                (interval, ignored1, ignored2, ignored3) -> interval.getEnd().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightStartMapping = interval -> interval.getStart().longValue();
+        Function<Interval<Long, BigDecimal>, Long> rightEndMapping = interval -> interval.getEnd().longValue();
+
+        AbstractPentaJoiner<Interval<Long, BigInteger>, String, String, String, Interval<Long, BigDecimal>> joiner =
+                (AbstractPentaJoiner<Interval<Long, BigInteger>, String, String, String, Interval<Long, BigDecimal>>) Joiners
+                        .overlapping(leftStartMapping,
+                                leftEndMapping, rightStartMapping, rightEndMapping);
+
+        assertSoftly(softly -> {
+            // True cases (equals, overlaps, contains, starts, ends)
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("Case A = B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 5L)))
+                    .as("B starts before A, A ends after B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A starts before B, B ends after A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 5L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B contains A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 5L)))
+                    .as("A contains B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("A started by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B started by A").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A ended by B").isTrue();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 7L)))
+                    .as("B ended by A").isTrue();
+
+            // False Cases (before, after, meets)
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(5L, 7L)))
+                    .as("A before B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(5L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B before A").isFalse();
+
+            // This is false since typically, when overlaps is used,
+            // end is exclusive, and start is inclusive,
+            // so 0-5, 5-10 do not overlap
+            softly.assertThat(joiner.matches(Interval.ofBigInt(1L, 3L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(3L, 7L)))
+                    .as("A meets B").isFalse();
+
+            softly.assertThat(joiner.matches(Interval.ofBigInt(3L, 7L),
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    "Ignored Arg",
+                    Interval.ofBigDecimal(1L, 3L)))
+                    .as("B meets A").isFalse();
+        });
+    }
+
 }
