+++ optaplanner-examples/src/main/java/org/optaplanner/examples/meetingscheduling/optional/score/MeetingSchedulingConstraintProvider.java
@@ -20,7 +20,7 @@ import static org.optaplanner.core.api.score.stream.Joiners.equal;
 import static org.optaplanner.core.api.score.stream.Joiners.filtering;
 import static org.optaplanner.core.api.score.stream.Joiners.greaterThan;
 import static org.optaplanner.core.api.score.stream.Joiners.lessThan;
-import static org.optaplanner.core.api.score.stream.Joiners.overlaps;
+import static org.optaplanner.core.api.score.stream.Joiners.overlapping;
 
 import org.optaplanner.core.api.score.stream.Constraint;
 import org.optaplanner.core.api.score.stream.ConstraintFactory;
@@ -62,7 +62,7 @@ public class MeetingSchedulingConstraintProvider implements ConstraintProvider {
                 .join(MeetingAssignment.class,
                         equal(MeetingAssignment::getRoom),
                         lessThan(MeetingAssignment::getId),
-                        overlaps(assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
+                        overlapping(assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
                                 assignment -> assignment.getStartingTimeGrain().getGrainIndex() +
                                         assignment.getMeeting().getDurationInGrains()))
                 .penalizeConfigurable("Room conflict",
@@ -89,7 +89,7 @@ public class MeetingSchedulingConstraintProvider implements ConstraintProvider {
                         equal((leftRequiredAttendance, rightRequiredAttendance, leftAssignment) -> rightRequiredAttendance
                                 .getMeeting(),
                                 MeetingAssignment::getMeeting),
-                        overlaps((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
+                        overlapping((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
                                 (attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex() +
                                         assignment.getMeeting().getDurationInGrains(),
                                 assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
@@ -133,7 +133,7 @@ public class MeetingSchedulingConstraintProvider implements ConstraintProvider {
                 .join(MeetingAssignment.class,
                         equal((requiredAttendance, preferredAttendance, leftAssignment) -> preferredAttendance.getMeeting(),
                                 MeetingAssignment::getMeeting),
-                        overlaps((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
+                        overlapping((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
                                 (attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex() +
                                         assignment.getMeeting().getDurationInGrains(),
                                 assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
@@ -153,7 +153,7 @@ public class MeetingSchedulingConstraintProvider implements ConstraintProvider {
                 .join(MeetingAssignment.class,
                         equal((leftAttendance, rightAttendance, leftAssignment) -> rightAttendance.getMeeting(),
                                 MeetingAssignment::getMeeting),
-                        overlaps((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
+                        overlapping((attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex(),
                                 (attendee1, attendee2, assignment) -> assignment.getStartingTimeGrain().getGrainIndex() +
                                         assignment.getMeeting().getDurationInGrains(),
                                 assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
@@ -189,7 +189,7 @@ public class MeetingSchedulingConstraintProvider implements ConstraintProvider {
                 .join(MeetingAssignment.class,
                         greaterThan((leftAssignment) -> leftAssignment.getMeeting().getId(),
                                 (rightAssignment) -> rightAssignment.getMeeting().getId()),
-                        overlaps(assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
+                        overlapping(assignment -> assignment.getStartingTimeGrain().getGrainIndex(),
                                 assignment -> assignment.getStartingTimeGrain().getGrainIndex() +
                                         assignment.getMeeting().getDurationInGrains()),
                         filtering((leftAssignment, rightAssignment) -> rightAssignment.getStartingTimeGrain() != null),
